Move button-related elements to the 'button' element group folder
