+++ angular/components/elements/element-group-button/confirm-button/confirm-button.component.scss
@@ -1,8 +1,8 @@
 // confirm-button.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 confirm-button {
     @include flex(center, center, row);
